Check for duplicates and enable multi-threading
There are cases where files in different folders have same name, but
different contents. Therefore, if a file was already copied but a new
one with the same name is found.  Check if they have the same content.
If same content, skip the copy.  Otherwise, append a number to the new
one and copy it to the destination.